zsbl: Split main.c into feature.c and jump.c by function
feature.c: setup_features
jump.c: Jump to the next stage entrance
