+++ Makefile
@@ -6,12 +6,12 @@ OBJCOPY     =   $(CROSS_COMPILE)objcopy
 OBJDUMP     =   $(CROSS_COMPILE)objdump
 CFLAGS = -fPIC
 TARGET = zero_stage_boot
-${TARGET}.elf: start.o main.o
+${TARGET}.elf: start.o feature.o jump.o
 	$(LD) -Tlink.lds $^ -o $@
 	$(OBJCOPY) -O binary $@ ${TARGET}.bin
 	$(OBJDUMP) -D $@ > ${TARGET}.asm
 
-main.o: main.c
+%.o: %.c
 	$(CC) -mcmodel=medany ${CFLAGS} -g -c $^
 
 start.o: start.S
